Rename module `feature_extractors` to `features`
