--- docs/source/api/morphoclass.feature_extractors.non_graph.rst
@@ -1,7 +1,0 @@
-morphoclass.feature\_extractors.non\_graph module
-=================================================
-
-.. automodule:: morphoclass.feature_extractors.non_graph
-   :members:
-   :undoc-members:
-   :show-inheritance:
