--- docs/source/api/morphoclass.feature_extractors.rst
@@ -1,18 +1,0 @@
-morphoclass.feature\_extractors package
-=======================================
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 4
-
-   morphoclass.feature_extractors.non_graph
-
-Module contents
----------------
-
-.. automodule:: morphoclass.feature_extractors
-   :members:
-   :undoc-members:
-   :show-inheritance:
