+++ docs/source/api/morphoclass.features.non_graph.rst
@@ -1,0 +1,7 @@
+morphoclass.features.non\_graph module
+======================================
+
+.. automodule:: morphoclass.features.non_graph
+   :members:
+   :undoc-members:
+   :show-inheritance:
