+++ docs/source/api/morphoclass.features.rst
@@ -1,0 +1,18 @@
+morphoclass.features package
+============================
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   morphoclass.features.non_graph
+
+Module contents
+---------------
+
+.. automodule:: morphoclass.features
+   :members:
+   :undoc-members:
+   :show-inheritance:
